Update stemming from video "II. Czech Classic"
Many new functions added: 1.) method to read more fields from a FIT file, including recorded distance and altitude. Both values can then be rendered as a new type of frames. 2.) The interpolation procedure was reworked, now using the methods from scipy. As a default, the Akima interpolator in the time domain is used. 3.) New, pretty much independent class was added - Stopwatch, which enables user to generate video frames of a countdown (used in videos of regularity tests, RTs). 4.) An attempt to paralelize the computation was made, using joblib library, but with only limited success. It seems to me that the PIL library is the actual bottleneck and it doesn't scale with the number of processes/threads using it.